backend/requirements.txt
# =============================================================================
# Backend Requirements - CodeMix NLP Platform
# Python 3.11+
# =============================================================================

# FastAPI Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.7
httpx==0.26.0

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2

# Redis
redis[asyncio]==5.0.1

# Celery
celery==5.3.6
kombu==5.3.4

# ML / Deep Learning
torch>=2.0.0
transformers==4.38.0
tokenizers>=0.15.0
accelerate==0.27.2
datasets==2.17.1
evaluate==0.4.1

# Explainability
shap==0.44.1
captum==0.7.0
scipy>=1.11.0

# Scientific
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Utils
python-dotenv==1.0.0
tabulate==0.9.0
tqdm>=4.65.0
einops==0.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
